Revisions and Updates
• New process flow of Treasury functions which include
1. Automated Check Voucher and number generation and printing. This
prints the check voucher with number continued from the existing system.
Then, to be submitted for manual signing of check.
2. Input of Check number and date into the system. This will send an
email notification to the employee for schedule of claiming. Treasury
has the option to cancel the check which will also cancel the
application in case of employee’s resignation or other reason of
cancellation.
3. Release of Check. Upon released of check, the system generates a
deduction schedule to be sent to the employee and to payroll. The
account is now available for maintenance of EFund custodian.
• Report generation of Treasury on Released Checks
• Notify payroll through email on Fully Paid Account.
• Approved EFund Form printing should not include the name of the
Officer but a check mark only